% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unemployment.R
\name{calc_arbmarkn_stat}
\alias{calc_arbmarkn_stat}
\title{Calculates labor force program statistics}
\usage{
calc_arbmarkn_stat(AEA_data)
}
\arguments{
\item{AEA_data}{AEA data set}
}
\description{
Calculates labor force program statistics
}

