% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_dik_stat}
\alias{calc_dik_stat}
\title{Calculate all statistics}
\usage{
calc_dik_stat(AEA_data)
}
\arguments{
\item{AEA_data}{Data set from AEA read into R.}
}
\description{
Calculates all statistics of interest and store it as list.
}

