% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unemployment.R
\name{calc_ers_stat}
\alias{calc_ers_stat}
\title{Calculates allowance statistics}
\usage{
calc_ers_stat(AEA_data)
}
\arguments{
\item{AEA_data}{AEA data set}
}
\description{
Calculate allowance statistics for the whole data.
}

