% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unemployment.R
\name{calc_member_aea_stat}
\alias{calc_member_aea_stat}
\title{Calculates membership statistics}
\usage{
calc_member_aea_stat(AEA_data)
}
\arguments{
\item{AEA_data}{AEA data set}
}
\description{
Calculate membership statistics
}

