% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_class}
\alias{check_class}
\title{Check if x has class AEA_data}
\usage{
check_class(x, class_name)
}
\arguments{
\item{x}{object to check.}

\item{class_name}{check for this class}
}
\description{
Test if x is an object of class AEA_data, otherwise generates warning.
}

