% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_dik_data.R
\name{check_dik_data}
\alias{check_dik_data}
\title{Control all the DIK data sets for errors}
\usage{
check_dik_data(db_path)
}
\arguments{
\item{db_path}{}
}
\description{
Control all the DIK data sets for errors
}
\details{
Run tests on the datasets in the DIK db to check for errors.
}

