% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DIK.R
\docType{package}
\name{DIK}
\alias{DIK}
\alias{DIK-package}
\title{DIK}
\description{
R package for analysis at DIK. Package contain specialized functions to handle
analysis tasks regarding unemployment statistics etc.
}

