% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dik_classify.R
\name{dik_classify}
\alias{dik_classify}
\title{Classifies based on DIK interest groups}
\usage{
dik_classify(x, type = c("utbildningsgrupp", "intressegrupp"),
  source_folder = NULL)
}
\arguments{
\item{x}{variable to classify}

\item{type}{which classification to use ("utbildningsgrupp" or "intressegrupp")}

\item{source_folder}{Folder that contain Utbildningsgrupper.csv and Intressegrupper.csv}
}
\description{
Classifies based on DIK interest groups
}
\details{
The functions classifies using the classifiers found here:
\url{https://github.com/MansMeg/DIK_analys/tree/master/Classification}
}

