% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_AEA_file.R
\name{find_stat_var_name}
\alias{find_stat_var_name}
\title{Automatically find out variable name}
\usage{
find_stat_var_name(x)
}
\arguments{
\item{x}{statX variable to check}
}
\description{
Automatically find out variable name
}

