% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_date}
\alias{to_date}
\title{Convert month and year to lubridate format}
\usage{
to_date(year, month, day = "01")
}
\arguments{
\item{year}{Year}

\item{month}{Month as text (ex. januari).}

\item{day}{Day to set (default is "01")}
}
\value{
vector in lubridate format
}
\description{
Creates a date variable
}

