% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_dik_stat}
\alias{write_dik_stat}
\title{Write statistics to csv-files}
\usage{
write_dik_stat(dik_stat, results_path)
}
\arguments{
\item{dik_stat}{A dik_stat object calculated with \code{calc_dik_stat}.}
}
\description{
Writes all statistics in a dik_stat object to different csv-files.
}

