% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggr_unempl_data}
\alias{aggr_unempl_data}
\title{Calculate all statistics}
\usage{
aggr_unempl_data(AEA_data)
}
\arguments{
\item{AEA_data}{Data set from AEA read into R.}
}
\description{
Calculates all statistics of interest and store it as list ready to append to data files.
}

