% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dik_classify}
\alias{dik_classify}
\title{Classifies based on DIK interest groups}
\usage{
dik_classify(x, type = c("utbildningsgrupp", "intressegrupp"))
}
\arguments{
\item{x}{variable to classify}

\item{type}{which classification to use ("utbildningsgrupp" or "intressegrupp")}
}
\description{
Classifies based on DIK interest groups
}
\details{
The functions classifies using the classifiers found here:
\url{https://github.com/MansMeg/DIK_analys/tree/master/Classification}
}

