% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_AEA_file}
\alias{read_AEA_file}
\title{Read in DIK unemployment files from AEA}
\usage{
read_AEA_file(file_path)
}
\arguments{
\item{file_path}{Path to an AEA file}
}
\value{
\code{data.frame} of class AEA_data.
}
\description{
All files are read in to R and put in one data.frame, checks that the file is correct
is conducted.
}

