% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_dik_stat_csv}
\alias{write_dik_stat_csv}
\title{Write statistics to csv-files}
\usage{
write_dik_stat_csv(dik_stat, results_path)
}
\arguments{
\item{dik_stat}{A dik_stat object calculated with \code{calc_dik_stat}.}
}
\description{
Writes all statistics in a dik_stat object to different csv-files.
}

